First Sleep Mode Implementation (UNO-noNano)
This time the sleep mode was tested with an Arduino UNO, instead of a
nano, because of the absence of a protoboard.
It is inportant to remark that I had to struggle a lot to find out a
problem which log was: "undefined reference to main" and "collect2.exe:
error: ld returned 5 exit status". The problem seemed to be that the
file was created with the Arduino IDE installe on C:/... but then I was
trying to compile it with the portable IDE.
The