First implementation of void readADXL
This function read the ADXL335 sensor 10 times. It takes about 1 second
to run the function, but its purpose is to avoid interference, so if
more than 5 times the mesure was't normal it means the hive is moving.
In the future the function is expected to return a boolean, instead of
the LED_BUILTIN HIGH/LOW, to tell the micro if it's necessary to send an
alarm to the beekeepr.