resetData() added, and loop develped to read sensors
The resetData function was created and invoked after sendingData.
There was created a constant value  dataInitialValue, which will not be
taken into account by the Main Unit when it receives it.
At same time, many functions were appended in the loop function, after
the goToSleep function. So when the micro wakes up, all the neccessary
data is messured and sended to the Main Unit.